Feat: 에러 바운더리 보완
Related to: #348
