+++ src/components/Main/MyChallenge/MyChallengeLinkWrap/MyChallengeLinkWrap.tsx
@@ -26,5 +26,4 @@ const MyChallengeLinkWrap = ({
     </div>
   );
 };
-
 export default MyChallengeLinkWrap;
