+++ src/components/Shop/ShopFrameList/ShopFrameList.tsx
@@ -28,7 +28,7 @@ const ShopFrameList = () => {
     if (getFrameItemSuccess) {
       onSuccessGetFrameItem();
     }
-  }, [getFrameItemSuccess, frameItemData]);
+  }, [getFrameItemSuccess]);
 
   return (
     <>
