+++ src/index.css
@@ -10,6 +10,10 @@ html {
   @apply font-SCoreDream;
 }
 
+#root {
+  height: 100vh;
+}
+
 :root {
   /* color-scheme: light dark; */
   font-synthesis: none;
