+++ src/pages/Certification/MyAllCurrentCertification/MyAllCurrentCertification.tsx
@@ -6,6 +6,9 @@ import { decrypt } from "@/hooks/useCrypto";
 import MyTotalCertification from "@/components/Certification/MyAllCurrentCertification/MyProfile/MyTotalCertification";
 import { Suspense } from "react";
 import LoadingBox from "@/components/Common/Loading/LoadingBox/LoadingBox";
+import { ErrorBoundary } from "react-error-boundary";
+import CommonGetErrorFallback from "@/components/Error/CommonGetErrorFallback/CommonGetErrorFallback";
+import { QueryErrorResetBoundary } from "react-query";
 
 function MyAllCurrentCertification() {
   const { id } = useParams();
@@ -17,15 +20,24 @@ function MyAllCurrentCertification() {
     <MobCard>
       <Header content="나의 인증 현황" />
       <div className="py-[6rem] px-[2.2rem] h-full">
-        <Suspense fallback={<LoadingBox />}>
-          <div className="mt-[3.4rem] _sm:mt-[1.8rem]">
-            <MyProfile decryptedUserId={decryptedUserId} />
-            <MyTotalCertification
-              decryptedInstanceId={decryptedInstanceId}
-              decryptedUserId={decryptedUserId}
-            />
-          </div>
-        </Suspense>
+        <QueryErrorResetBoundary>
+          {({ reset }) => (
+            <ErrorBoundary
+              onReset={reset}
+              FallbackComponent={CommonGetErrorFallback}
+            >
+              <Suspense fallback={<LoadingBox />}>
+                <div className="mt-[3.4rem] _sm:mt-[1.8rem]">
+                  <MyProfile decryptedUserId={decryptedUserId} />
+                  <MyTotalCertification
+                    decryptedInstanceId={decryptedInstanceId}
+                    decryptedUserId={decryptedUserId}
+                  />
+                </div>
+              </Suspense>
+            </ErrorBoundary>
+          )}
+        </QueryErrorResetBoundary>
       </div>
     </MobCard>
   );
