+++ src/pages/Home/NewChallenge/NewChallenge.tsx
@@ -1,39 +1,24 @@
-import { useEffect, useState } from "react";
-import VerticalChallengeItems from "@/components/Common/VerticalChallengeItems/VerticalChallengeItems";
 import HomeLayout from "@/layout/HomeLayout/HomeLayout";
-import { useInView } from "react-intersection-observer";
-import LoadingBox from "@/components/Common/Loading/LoadingBox/LoadingBox";
-import { useGetInfiniteLastestInstance } from "@/hooks/queries/useHomeInstanceQuery";
-import { InstanceThumbnailDataType } from "@/types/homeInstance";
+import { ErrorBoundary } from "react-error-boundary";
+import CommonGetErrorFallback from "@/components/Error/CommonGetErrorFallback/CommonGetErrorFallback";
+import { QueryErrorResetBoundary } from "react-query";
+import InfiniteNewChallenge from "@/components/Home/InfiniteNewChallenge/InfiniteNewChallenge";
 
 const NewChallenge = () => {
-  const [ref, inView] = useInView();
-  const [challenges, setChallenges] = useState<InstanceThumbnailDataType[]>([]);
-
-  const { fetchNextPage, hasNextPage, isLoading } =
-    useGetInfiniteLastestInstance({ setChallenges });
-
-  useEffect(() => {
-    if (inView && hasNextPage) {
-      fetchNextPage();
-    }
-  }, [inView, hasNextPage, fetchNextPage]);
-
-  if (!challenges) return null;
-
   return (
     <HomeLayout>
-      {isLoading ? (
-        <LoadingBox />
-      ) : (
-        <div className="mx-[2.2rem] mt-[1rem]">
-          <VerticalChallengeItems data={challenges} />
-          <div
-            ref={ref}
-            style={{ height: "10px", background: "transparent" }}
-          />
-        </div>
-      )}
+      <div className="mx-[2.2rem] mt-[1rem] h-full">
+        <QueryErrorResetBoundary>
+          {({ reset }) => (
+            <ErrorBoundary
+              onReset={reset}
+              FallbackComponent={CommonGetErrorFallback}
+            >
+              <InfiniteNewChallenge />
+            </ErrorBoundary>
+          )}
+        </QueryErrorResetBoundary>
+      </div>
     </HomeLayout>
   );
 };
