+++ src/pages/Home/Search/ActivitySearch/ActivitySearch.tsx
@@ -1,9 +1,7 @@
-import LoadingBox from "@/components/Common/Loading/LoadingBox/LoadingBox";
-import VerticalChallengeItems from "@/components/Common/VerticalChallengeItems/VerticalChallengeItems";
-import { useGetSearchInfiniteActivityInstance } from "@/hooks/queries/useHomeInstanceQuery";
-import { InstanceThumbnailDataType } from "@/types/homeInstance";
-import { useEffect, useState } from "react";
-import { useInView } from "react-intersection-observer";
+import CommonGetErrorFallback from "@/components/Error/CommonGetErrorFallback/CommonGetErrorFallback";
+import InfiniteActivitySearchedChallenge from "@/components/Home/InfiniteActivitySearchedChallenge/InfiniteActivitySearchedChallenge";
+import { ErrorBoundary } from "react-error-boundary";
+import { QueryErrorResetBoundary } from "react-query";
 import { useOutletContext } from "react-router-dom";
 
 interface Outlet {
@@ -16,32 +14,22 @@ function ActivitySearch() {
   const { searchQuery, searchEnter, setSearchEnter } =
     useOutletContext<Outlet>();
 
-  const [ref, inView] = useInView();
-  const [challenges, setChallenges] = useState<InstanceThumbnailDataType[]>([]);
-
-  const { fetchNextPage, hasNextPage, refetch, isLoading } =
-    useGetSearchInfiniteActivityInstance({ searchQuery, setChallenges });
-
-  useEffect(() => {
-    if (inView && hasNextPage) {
-      fetchNextPage();
-    }
-  }, [inView, hasNextPage, fetchNextPage]);
-
-  useEffect(() => {
-    if (searchEnter) {
-      refetch();
-      setSearchEnter(false);
-    }
-  }, [searchEnter]);
-
-  if (!challenges) return null;
-  if (isLoading) return <LoadingBox />;
-
   return (
     <>
-      <VerticalChallengeItems data={challenges} />
-      <div ref={ref} style={{ height: "10px" }}></div>
+      <QueryErrorResetBoundary>
+        {({ reset }) => (
+          <ErrorBoundary
+            onReset={reset}
+            FallbackComponent={CommonGetErrorFallback}
+          >
+            <InfiniteActivitySearchedChallenge
+              searchEnter={searchEnter}
+              searchQuery={searchQuery}
+              setSearchEnter={setSearchEnter}
+            />
+          </ErrorBoundary>
+        )}
+      </QueryErrorResetBoundary>
     </>
   );
 }
