+++ src/pages/Home/Search/AllSearch/AllSearch.tsx
@@ -1,10 +1,8 @@
-import LoadingBox from "@/components/Common/Loading/LoadingBox/LoadingBox";
-import VerticalChallengeItems from "@/components/Common/VerticalChallengeItems/VerticalChallengeItems";
-import { useGetSearchInfiniteAllInstance } from "@/hooks/queries/useHomeInstanceQuery";
-import { useEffect, useState } from "react";
-import { useInView } from "react-intersection-observer";
 import { useOutletContext } from "react-router-dom";
-import { InstanceThumbnailDataType } from "@/types/homeInstance";
+import { ErrorBoundary } from "react-error-boundary";
+import { QueryErrorResetBoundary } from "react-query";
+import CommonGetErrorFallback from "@/components/Error/CommonGetErrorFallback/CommonGetErrorFallback";
+import InfiniteAllSearchedChallenge from "@/components/Home/InfiniteAllSearchedChallenge/InfiniteAllSearchedChallenge";
 
 interface Outlet {
   searchQuery: string;
@@ -16,32 +14,22 @@ function AllSearch() {
   const { searchQuery, searchEnter, setSearchEnter } =
     useOutletContext<Outlet>();
 
-  const [ref, inView] = useInView();
-  const [challenges, setChallenges] = useState<InstanceThumbnailDataType[]>([]);
-
-  const { fetchNextPage, hasNextPage, refetch, isLoading } =
-    useGetSearchInfiniteAllInstance({ setChallenges, searchQuery });
-
-  useEffect(() => {
-    if (inView && hasNextPage) {
-      fetchNextPage();
-    }
-  }, [inView, hasNextPage, fetchNextPage]);
-
-  useEffect(() => {
-    if (searchEnter) {
-      refetch();
-      setSearchEnter(false);
-    }
-  }, [searchEnter]);
-
-  if (!challenges) return null;
-  if (isLoading) return <LoadingBox />;
-
   return (
     <>
-      <VerticalChallengeItems data={challenges} />
-      <div ref={ref} style={{ height: "10px" }}></div>
+      <QueryErrorResetBoundary>
+        {({ reset }) => (
+          <ErrorBoundary
+            onReset={reset}
+            FallbackComponent={CommonGetErrorFallback}
+          >
+            <InfiniteAllSearchedChallenge
+              searchEnter={searchEnter}
+              searchQuery={searchQuery}
+              setSearchEnter={setSearchEnter}
+            />
+          </ErrorBoundary>
+        )}
+      </QueryErrorResetBoundary>
     </>
   );
 }
