+++ src/pages/Home/Search/DoneSearch/DoneSearch.tsx
@@ -1,10 +1,8 @@
-import LoadingBox from "@/components/Common/Loading/LoadingBox/LoadingBox";
-import VerticalChallengeItems from "@/components/Common/VerticalChallengeItems/VerticalChallengeItems";
-import { useGetSearchInfiniteDoneInstance } from "@/hooks/queries/useHomeInstanceQuery";
-import { useEffect, useState } from "react";
-import { useInView } from "react-intersection-observer";
 import { useOutletContext } from "react-router-dom";
-import { InstanceThumbnailDataType } from "@/types/homeInstance";
+import { QueryErrorResetBoundary } from "react-query";
+import { ErrorBoundary } from "react-error-boundary";
+import CommonGetErrorFallback from "@/components/Error/CommonGetErrorFallback/CommonGetErrorFallback";
+import InfiniteDoneSearchedChallenge from "@/components/Home/InfiniteDoneSearchedChallenge/InfiniteDoneSearchedChallenge";
 
 interface Outlet {
   searchQuery: string;
@@ -16,31 +14,22 @@ function DoneSearch() {
   const { searchQuery, searchEnter, setSearchEnter } =
     useOutletContext<Outlet>();
 
-  const [ref, inView] = useInView();
-  const [challenges, setChallenges] = useState<InstanceThumbnailDataType[]>([]);
-
-  const { fetchNextPage, hasNextPage, refetch, isLoading } =
-    useGetSearchInfiniteDoneInstance({ searchQuery, setChallenges });
-
-  useEffect(() => {
-    if (inView && hasNextPage) {
-      fetchNextPage();
-    }
-  }, [inView, hasNextPage, fetchNextPage]);
-
-  useEffect(() => {
-    if (searchEnter) {
-      refetch();
-      setSearchEnter(false);
-    }
-  }, [searchEnter]);
-
-  if (!challenges) return null;
-  if (isLoading) return <LoadingBox />;
   return (
     <>
-      <VerticalChallengeItems data={challenges} />
-      <div ref={ref} style={{ height: "10px" }}></div>
+      <QueryErrorResetBoundary>
+        {({ reset }) => (
+          <ErrorBoundary
+            onReset={reset}
+            FallbackComponent={CommonGetErrorFallback}
+          >
+            <InfiniteDoneSearchedChallenge
+              searchEnter={searchEnter}
+              searchQuery={searchQuery}
+              setSearchEnter={setSearchEnter}
+            />
+          </ErrorBoundary>
+        )}
+      </QueryErrorResetBoundary>
     </>
   );
 }
