+++ src/pages/Home/Search/PreActivitySearch/PreActivitySearch.tsx
@@ -1,10 +1,8 @@
-import LoadingBox from "@/components/Common/Loading/LoadingBox/LoadingBox";
-import VerticalChallengeItems from "@/components/Common/VerticalChallengeItems/VerticalChallengeItems";
-import { useGetSearchInfinitePreActivityInstance } from "@/hooks/queries/useHomeInstanceQuery";
-import { useEffect, useState } from "react";
-import { useInView } from "react-intersection-observer";
 import { useOutletContext } from "react-router-dom";
-import { InstanceThumbnailDataType } from "@/types/homeInstance";
+import { QueryErrorResetBoundary } from "react-query";
+import { ErrorBoundary } from "react-error-boundary";
+import CommonGetErrorFallback from "@/components/Error/CommonGetErrorFallback/CommonGetErrorFallback";
+import InfinitePreActivitySearchedChallenge from "@/components/Home/InfinitePreActivitySearchedChallenge/InfinitePreActivitySearchedChallenge";
 
 interface Outlet {
   searchQuery: string;
@@ -16,31 +14,22 @@ function PreActivitySearch() {
   const { searchQuery, searchEnter, setSearchEnter } =
     useOutletContext<Outlet>();
 
-  const [ref, inView] = useInView();
-  const [challenges, setChallenges] = useState<InstanceThumbnailDataType[]>([]);
-
-  const { fetchNextPage, hasNextPage, refetch, isLoading } =
-    useGetSearchInfinitePreActivityInstance({ setChallenges, searchQuery });
-
-  useEffect(() => {
-    if (inView && hasNextPage) {
-      fetchNextPage();
-    }
-  }, [inView, hasNextPage, fetchNextPage]);
-
-  useEffect(() => {
-    if (searchEnter) {
-      refetch();
-      setSearchEnter(false);
-    }
-  }, [searchEnter]);
-
-  if (!challenges) return null;
-  if (isLoading) return <LoadingBox />;
   return (
     <>
-      <VerticalChallengeItems data={challenges} />
-      <div ref={ref} style={{ height: "10px" }}></div>
+      <QueryErrorResetBoundary>
+        {({ reset }) => (
+          <ErrorBoundary
+            onReset={reset}
+            FallbackComponent={CommonGetErrorFallback}
+          >
+            <InfinitePreActivitySearchedChallenge
+              searchEnter={searchEnter}
+              searchQuery={searchQuery}
+              setSearchEnter={setSearchEnter}
+            />
+          </ErrorBoundary>
+        )}
+      </QueryErrorResetBoundary>
     </>
   );
 }
