--- src/pages/MyChallenge/MyChallengeStyle.css
@@ -1,3 +1,0 @@
-.myChallenge-wrap {
-  @apply w-full max-w-[48.6rem];
-}
